+++ apis/vara-eth/src/types/api/program.ts
@@ -1,30 +1,11 @@
-import { HexString } from '../../types/index.js';
+import type { Hex } from 'viem';
+
+import type { ReplyCode } from '../../errors/index.js';
 
 export interface ReplyInfo {
-  readonly payload: HexString;
+  readonly payload: Hex;
   readonly value: number;
   readonly code: ReplyCode;
-}
-
-export type ReplyCode = SuccessReplyCode | ErrorReplyCode;
-
-interface SuccessReplyCode {
-  readonly Success: 'Manual' | 'Auto';
-}
-
-interface ErrorReplyCode {
-  readonly Error: ExecutionReason;
-}
-
-interface ExecutionReason {
-  readonly Execution:
-    | 'RanOutOfGas'
-    | 'MemoryOverflow'
-    | 'BackendError'
-    | 'UserspacePanic'
-    | 'UnreachableInstruction'
-    | 'StackLimitExceeded'
-    | 'Unsupported';
 }
 
 export interface ProgramState {