+++ idea/vara-eth/frontend/src/features/codes/lib/requests.ts
@@ -1,12 +1,12 @@
-import { HexString } from '@vara-eth/api';
+import { Hex } from 'viem';
 
 import { EXPLORER_URL } from '@/shared/config';
 import { fetchWithGuard } from '@/shared/utils';
 
 export type Code = {
-  id: HexString;
+  id: Hex;
   status: string;
   createdAt: string;
 };
 
-export const getCode = (id: HexString) => fetchWithGuard<Code>({ url: `${EXPLORER_URL}/codes/${id}` });
+export const getCode = (id: Hex) => fetchWithGuard<Code>({ url: `${EXPLORER_URL}/codes/${id}` });