+++ idea/vara-eth/frontend/src/features/search/lib/requests.ts
@@ -1,4 +1,4 @@
-import { HexString } from '@vara-eth/api';
+import { Hex } from 'viem';
 
 import { EXPLORER_URL } from '@/shared/config';
 import { fetchWithGuard } from '@/shared/utils';
@@ -18,5 +18,5 @@ export const INDEXER_ENTITY = {
 
 type IndexerEntityResponse = { type: (typeof INDEXER_ENTITY)[keyof typeof INDEXER_ENTITY]; data: unknown };
 
-export const getIndexerEntity = (hash: HexString) =>
+export const getIndexerEntity = (hash: Hex) =>
   fetchWithGuard<IndexerEntityResponse>({ url: `${EXPLORER_URL}/lookup/${hash}` });